Update: Sale discount column

+++ database/migrations/2025_05_11_225429_create_sales_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('sales', function (Blueprint $table) {
             $table->id();
             $table->decimal('subtotal', 8, 2)->default(0);
-            $table->decimal('discount',8,2)->default(0);
+            $table->unsignedTinyInteger('discount')->default(0);
             $table->decimal('total_amount',8,2)->default(0);
             $table->enum('order_type', ['dine-in', 'take-out'])->default('dine-in');
             $table->enum('payment_method', ['cash', 'gcash'])->default('cash');
